+++ lib/src/existing_steps.dart
@@ -2,27 +2,34 @@ import 'dart:io';
 
 import 'package:path/path.dart' as p;
 
+/// key - step filename, value - path for import (ex: {'i_have_a_step.dart': 'step/common'})
 Map<String, String> getExistingStepSubfolders(
-    String featureDir, String stepFolderName) {
+  String featureDir,
+  String stepFolderName,
+) {
   final stepFolder = p.join(featureDir, stepFolderName);
   final steps = Directory(stepFolder);
   if (!steps.existsSync()) {
     return {};
   }
   return steps.listSync(recursive: true).asMap().map(
-        (key, value) => MapEntry<String, String>(
-          p.basename(value.path),
-          _getStepFolder(
-            value.uri.pathSegments,
+        (_, step) => MapEntry<String, String>(
+          p.basename(step.path),
+          _getStepSubfolders(
+            steps.uri.pathSegments.length,
+            step.uri.pathSegments,
             stepFolderName,
           ),
         ),
       );
 }
 
-String _getStepFolder(List<String> pathSegments, String stepFolderName) {
-  final stepFolderIndex = pathSegments.indexOf(stepFolderName);
-  final stepFolder =
-      pathSegments.getRange(stepFolderIndex, pathSegments.length - 1);
-  return p.joinAll(stepFolder);
+String _getStepSubfolders(
+  int stepFolderPathSegmentsLength,
+  List<String> currentStepPath,
+  String stepFolderName,
+) {
+  final pathDiff = currentStepPath.getRange(
+      stepFolderPathSegmentsLength - 1, currentStepPath.length - 1);
+  return p.joinAll([stepFolderName, ...pathDiff]);
 }