+++ test/step_folder_name_test.dart
@@ -17,7 +17,7 @@ Feature: Testing feature
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import './custom_steps/the_app_is_running.dart';
+import './../../../custom_steps/the_app_is_running.dart';
 
 void main() {
   group('Testing feature', () {
@@ -32,7 +32,8 @@ void main() {
       featureDir: 'test.feature',
       package: 'test',
       input: featureFile,
-      generatorOptions: const GeneratorOptions(stepFolderName: 'custom_steps'),
+      generatorOptions:
+          const GeneratorOptions(stepFolderName: '../../../custom_steps'),
     );
     expect(feature.dartContent, expectedFeatureDart);
   });